Trade IDs, goods rem. +handling impr., adjust endpoints, no locked goods

+++ src/main/java/net/teamio/gtams/server/info/Trade.java
@@ -6,7 +6,7 @@ public class Trade {
 
 	public UUID terminalId;
 
-	public int tradeId;
+	public long tradeId;
 
 	public TradeDescriptor descriptor;
